Don't add to 'all' team unless actually necessary
This rewrite does better at avoiding the awful race-condition where we
might add someone to the 'all' team immediately after they've been
removed from the Org. We check org-membership /after/ we've established
they're not in the 'all' team, and then add the user immediately, only if
they /are/ a member of the Org.

Note that membership of the 'all' team implies membership of the Org, so
no need to do the `user.isMemberOf(org)` call if the user is already a
member of the team.

https://github.com/kohsuke/github-api/pull/75