Use OkHttp for it's HttpResponseCache/DiskLruCache
This mitigates well against the rate-limiting used by the GitHub API -
they welcome conditional requests and so calls for unchanged content
don't cost us. Using caching takes rate-limit cost from 320 to 106
for a run against the Guardian org- total allowance for an hour is 5000.
Even with an organisation 5x the size, gu-who should be fine to run every
5 minutes.

http://developer.github.com/v3/#conditional-requests
https://github.com/kohsuke/github-api/pull/74