Cope with deleted users - user accounts which are literally gone
If a user account gets shutdown, then the issue assigned to that account
is no longer assigned to them - the assignee field is blank. At this point,
gu-who now just closes the issue, as there's nothing further we can do
(we don't want to bother trying to parse the issue headline, BORING!).