Tolerate inability to evaluate some requirements, eg Two-Factor-Auth
The TFA-auditing API call is only available for owners of organizations
with private repositories (odd- don't you want to check the people who
can write to public repos?!) so it will fail for any non-paid-for GitHub
org. This commit gives the ability to tolerate that call failure, by
introducing the AccountRequirement UserEvaluator, which can only be
'successfully' instantiated if the required data is available, but uses
the 'Try' class to capture the failure nicely.

https://developer.github.com/v3/orgs/members/#members-list