Better logging and behaviour when no users.txt file found (issue #21)
The users.txt file is a list of 'sponsored' users - for each new user
joining the GitHub organisation, a 'more-senior' member of staff adds the
user the users.txt file in 'people' repo, taking responsibility via
git-blame for the user being in the organisation. This ensures there’s
always someone to go to when membership for a dodgy account is in doubt.

gu:who always expects the file to be in the people repo, and behaved badly
when it couldn't find it:

```
[info] + extract sponsored user logins after cloning a remote repo
[info] ! fail when no users.txt found
[error]    NullPointerException:   (UnpackedObjectCache.java:146)
[error] org.eclipse.jgit.internal.storage.file.UnpackedObjectCache$Table.index(UnpackedObjectCache.java:146)
[error] org.eclipse.jgit.internal.storage.file.UnpackedObjectCache$Table.contains(UnpackedObjectCache.java:109)
[error] org.eclipse.jgit.internal.storage.file.UnpackedObjectCache.isUnpacked(UnpackedObjectCache.java:64)
[error] org.eclipse.jgit.internal.storage.file.ObjectDirectory.openObject(ObjectDirectory.java:367)
[error] org.eclipse.jgit.internal.storage.file.WindowCursor.open(WindowCursor.java:145)
[error] org.eclipse.jgit.lib.ObjectReader.open(ObjectReader.java:229)
[error] com.madgag.git.package$RichObjectId.open(package.scala:185)
[error] lib.PeopleRepo$$anonfun$2.apply(PeopleRepo.scala:65)
[error] lib.PeopleRepo$$anonfun$2.apply(PeopleRepo.scala:65)
```

It's now more tolerant of the missing file - it just means /no-one/ is
sponsored - and logs a warning.

Thanks to @timmattison for reporting the bad behaviour!